CI/CD설정 변경 및 오류 수정

+++ back-end/plogging/src/main/java/com/backend/plogging/dto/request/plogging/PathRequestDto.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PathRequestDto {
-    private Float lat;
-    private Float lng;
+    private Double lat;
+    private Double lng;
 }
